updated folder strucutre
