--- iot-e2e-tests/android-iot-device-tests/app/build.gradle
@@ -1,64 +1,0 @@
-apply plugin: 'com.android.application'
-
-android {
-    compileSdkVersion 23
-    buildToolsVersion '25.0.0'
-
-    defaultConfig {
-        applicationId "com.iothub.azure.microsoft.com.androidsample"
-        minSdkVersion 23
-        targetSdkVersion 23
-        versionCode 1
-        versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-        debug {
-            debuggable true
-        }
-    }
-
-    packagingOptions {
-        exclude "META-INF/*.SF"
-        exclude "META-INF/*.DSA"
-        exclude "META-INF/*.RSA"
-        exclude 'META-INF/DEPENDENCIES'
-        exclude 'META-INF/NOTICE'
-        exclude 'META-INF/LICENSE'
-        exclude 'META-INF/LICENSE.txt'
-        exclude 'META-INF/NOTICE.txt'
-        exclude 'thirdpartynotice.txt'
-    }
-    dexOptions {
-        incremental true
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
-    }
-}
-
-dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    testCompile 'junit:junit:4.12'
-    compile 'com.android.support:appcompat-v7:23.3.0'
-    androidTestCompile 'com.android.support:support-annotations:23.3.0'
-    androidTestCompile 'com.android.support.test:runner:0.5'
-    androidTestCompile 'junit:junit:4.12'
-    compile files('lib/junit-common-code.jar')
-    compile files('lib/iot-device-client-1.3.30-with-deps.jar')
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    // Remote binary dependency
-    //compile 'com.microsoft.azure.sdk.iot:iot-device-client:1.1.25'
-    //compile 'com.microsoft.azure.sdk.iot:iot-service-client:1.7.23'
-}
