+++ iot-e2e-tests/android/app/proguard-rules.pro
