+++ iot-e2e-tests/android/app/src/main/java/com/iothub/azure/microsoft/com/android/MainActivity.java
