+++ iot-e2e-tests/android/app/src/main/res/layout/activity_main.xml
