+++ iot-e2e-tests/android/app/src/main/res/mipmap-hdpi/ic_launcher.png
