+++ iot-e2e-tests/android/app/src/main/res/mipmap-mdpi/ic_launcher.png
