+++ iot-e2e-tests/android/app/src/main/res/mipmap-xhdpi/ic_launcher.png
