+++ iot-e2e-tests/android/app/src/main/res/mipmap-xxhdpi/ic_launcher.png
