+++ iot-e2e-tests/android/app/src/main/res/mipmap-xxxhdpi/ic_launcher.png
