+++ iot-e2e-tests/android/app/src/main/res/values-w820dp/dimens.xml
