+++ iot-e2e-tests/android/app/src/main/res/values/strings.xml
