+++ iot-e2e-tests/android/gradle.properties
