+++ iot-e2e-tests/android/gradle/wrapper/gradle-wrapper.properties
