+++ iot-e2e-tests/android/local.properties
