+++ iot-e2e-tests/android/settings.gradle
