--- iot-e2e-tests/common-test-code/src/tests/integration/com/microsoft/azure/sdk/iot/common/Success.java
@@ -1,19 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-package tests.integration.com.microsoft.azure.sdk.iot.common;
-
-public class Success
-{
-    public Boolean result = false;
-
-    public void setResult(Boolean result)
-    {
-        this.result = result;
-    }
-
-    public Boolean getResult()
-    {
-        return this.result;
-    }
-}
