+++ iot-e2e-tests/jvm/pom.xml
@@ -1,12 +1,14 @@
-<!-- Copyright (c) Microsoft. All rights reserved. --><!-- Licensed under the MIT license. See LICENSE file in the project root for full license information. -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<!--
+  ~  Copyright (c) Microsoft. All rights reserved.
+  ~  Licensed under the MIT license. See LICENSE file in the project root for full license information.
+  --><!-- Licensed under the MIT license. See LICENSE file in the project root for full license information. -->
+<project>
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.microsoft.azure.sdk.iot</groupId>
-    <artifactId>iot-device-client</artifactId>
-    <name>IoT Hub Java Device Client</name>
-    <version>1.5.33</version>
-    <description>The Microsoft Azure IoT Device SDK for Java</description>
+    <artifactId>iot-e2e-jvm-tests</artifactId>
+    <name>IoT Hub Java SDK jvm tests</name>
+    <version>0.9.13</version>
+    <description>Test suite fot the Microsoft Azure IoT Device SDK for Java</description>
     <developers>
         <developer>
             <id>microsoft</id>
@@ -15,28 +17,19 @@
     </developers>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <skipITs>true</skipITs>
     </properties>
     <dependencies>
         <dependency>
-            <groupId>org.apache.qpid</groupId>
-            <artifactId>proton-j</artifactId>
-            <version>0.18.0</version>
-            <type>jar</type>
+            <groupId>com.microsoft.azure.sdk.iot</groupId>
+            <artifactId>iot-device-client</artifactId>
+            <version>1.3.33</version>
         </dependency>
         <dependency>
-        <groupId>com.microsoft.azure.sdk.iot</groupId>
-        <artifactId>iot-deps</artifactId>
-        <version>0.1.2</version>
-        </dependency>
-        <dependency>
-            <groupId>com.microsoft.azure</groupId>
-            <artifactId>azure-storage</artifactId>
-            <version>2.2.0</version>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.paho</groupId>
-            <artifactId>org.eclipse.paho.client.mqttv3</artifactId>
-            <version>1.1.0</version>
+            <groupId>com.microsoft.azure.sdk.iot</groupId>
+            <artifactId>iot-service-client</artifactId>
+            <version>1.8.24</version>
+            <scope>test</scope>
         </dependency>
         <!-- test dependencies -->
         <dependency>
@@ -57,21 +50,10 @@
             <version>1.3</version>
             <scope>test</scope>
         </dependency>
-        <!-- https://mvnrepository.com/artifact/com.google.code.gson/gson -->
         <dependency>
-            <groupId>com.google.code.gson</groupId>
-            <artifactId>gson</artifactId>
-            <version>2.8.0</version>
-        </dependency>
-        <dependency>
-            <groupId>commons-codec</groupId>
-            <artifactId>commons-codec</artifactId>
-            <version>1.10</version>
-        </dependency>
-        <dependency>
-            <groupId>log4j</groupId>
-            <artifactId>log4j</artifactId>
-            <version>1.2.17</version>
+            <groupId>com.microsoft.azure.sdk.iot</groupId>
+            <artifactId>iot-e2e-common</artifactId>
+            <version>0.9.13</version>
         </dependency>
     </dependencies>
     <build>
@@ -112,6 +94,7 @@
                 <configuration>
                     <forkCount>1</forkCount>
                     <reuseForks>true</reuseForks>
+                    <skipITs>${skipITs}</skipITs>
                     <argLine>
                         -javaagent:${settings.localRepository}/org/jmockit/jmockit/1.22/jmockit-1.22.jar
                     </argLine>
