+++ iot-e2e-tests/jvm/src/test/java/tests/integration/com/microsoft/azure/sdk/iot/net/SignatureIT.java
