+++ iot-e2e-tests/pom.xml
@@ -1,0 +1,34 @@
+<!-- Copyright (c) Microsoft. All rights reserved. -->
+<!-- Licensed under the MIT license. See LICENSE file in the project root for full license information. -->
+<project>
+    <modelVersion>4.0.0</modelVersion>
+    <groupId>com.microsoft.azure.sdk.iot</groupId>
+	<artifactId>iot-e2e-tests</artifactId>
+	<name>IoT Hub Java SDK tests</name>
+	<version>0.9.13</version>
+	<description>Test suite fot the Microsoft Azure IoT Device SDK for Java</description> 
+    <packaging>pom</packaging>
+    <developers>
+        <developer>
+            <id>microsoft</id>
+            <name>Microsoft</name>
+        </developer>
+    </developers>
+    <modules>
+        <module>common</module>
+        <module>jvm</module>
+    </modules>
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.3</version>
+                <configuration>
+                    <source>1.8</source>
+                    <target>1.8</target>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
+</project>
